+++ tests/ContentSecurityPolicy/Violation/ReportEventTest.php
@@ -11,19 +11,19 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Nelmio\SecurityBundle\Tests\ContentSecurityPolicy\Violation;
+namespace ContentSecurityPolicy\Violation;
 
-use Nelmio\SecurityBundle\ContentSecurityPolicy\Violation\Event;
 use Nelmio\SecurityBundle\ContentSecurityPolicy\Violation\Report;
-use PHPUnit\Framework\TestCase;
+use Nelmio\SecurityBundle\ContentSecurityPolicy\Violation\ReportEvent;
+use PhpUnit\Framework\TestCase;
 
-class EventTest extends TestCase
+class ReportEventTest extends TestCase
 {
-    public function testCanBeConstructed(): void
+    public function testCanBeConstructed()
     {
         $this->assertInstanceOf(
-            Event::class,
-            new Event(new Report([]))
+            ReportEvent::class,
+            new ReportEvent(new Report([]))
         );
     }
 }