+++ CHANGELOG.md
@@ -5,9 +5,13 @@
 * Removed `DoctrineCacheUAFamilyParser` (use `PsrCacheUAFamilyParser` instead)
 * All classes have been marked as `final`
 
-### 2.11.1 (xxxx-xx-xx)
+### 2.12.0 (xxxx-xx-xx)
 * Filter moz-extension reports
 * Log user agent along with CSP report
+* Deprecated external_redirects.whitelist option in favor of external_redirects.allow_list
+* Deprecated forced_ssl.whitelist option in favor of forced_ssl.allow_list
+* Deprecated `Nelmio\SecurityBundle\ContentSecurityPolicy\Violation\Event` class in favor of
+  `Nelmio\SecurityBundle\ContentSecurityPolicy\Violation\ReportEvent`.
 
 ### 2.11.0 (2022-01-18)
   * Added support for CSP "prefetch-src" directive